requirements.txt
django
djangorestframework
pandas
django-dotenv
mysqlclient
django-cors-headers
requests